:recycle: 결제 관련 리팩토링

+++ src/main/java/kr/hhplus/concert/presentation/dto/ReservationDTO.java
@@ -15,8 +15,8 @@ public class ReservationDTO {
     public record ReservationRequest (
         Long userId,
         Long concertId,
-        Integer concertScheduleId,
-        Integer seatId
+        Long concertScheduleId,
+        Long seatId
     ) {
         public ReservationCommand toCommand(String token) {
             return ReservationCommand.builder()
